# Conflicts:
#	web Drinky-Winky/src/main/java/com/infoshare/servlet/SearchFavouriteServlet.java
#	web Drinky-Winky/src/main/webapp/WEB-INF/template/start.ftlh
#	web Drinky-Winky/src/main/webapp/WEB-INF/template/view.ftlh

+++ web Drinky-Winky/src/main/java/com/infoshare/service/DrinkService.java
@@ -11,6 +11,7 @@ import javax.enterprise.context.RequestScoped;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @RequestScoped
@@ -20,9 +21,13 @@ public class DrinkService {
     @EJB
     DrinkDao drinkDao;
 
-    public void addDrink(DrinkDTO drinkDTO) {
-        Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
-        drinkDao.addDrink(drink);
+    public String addDrink(DrinkDTO drinkDTO) {
+        if (drinkDao.getDrinkList().stream().noneMatch(drink -> drink.getName().equals(drinkDTO.getName()))) {
+            Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
+            drinkDao.addDrink(drink);
+            return "SUCCESS! THE DRINK HAS BEEN ADDED";
+        }
+        return "SOMETHING WENT WRONG! THE DRINK WITH THAT NAME ALREADY EXIST";
     }
 
     public void editDrink(DrinkDTO drinkDTO) {
@@ -50,14 +55,18 @@ public class DrinkService {
 
     @Transactional
     public void deleteDrinkById(Long id) {
-        drinkDao.deleteDrinkById(id);
-        logger.info("Drink has been deleted");
+        if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getId().equals(id))) {
+            drinkDao.deleteDrinkById(id);
+            logger.info("Drink has been deleted");
+        }
     }
 
     @Transactional
     public void deleteDrinkByName(String name) {
-        drinkDao.deleteDrinkByName(name);
-        logger.info("Drink has been deleted");
+        if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getName().equals(name))) {
+            drinkDao.deleteDrinkByName(name);
+            logger.info("Drink has been deleted");
+        }
     }
 
     @Transactional
@@ -66,6 +75,31 @@ public class DrinkService {
                 .stream()
                 .map(DrinkDTO::drinkToDTO)
                 .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public List<DrinkDTO> getRequestDrinkList(int request, int size) {
+
+        int fromIndex = (request - 1) * size;
+        int toIndex = request * size;
+
+        if (toIndex > drinkDao.getDrinkList().size()) {
+            toIndex = drinkDao.getDrinkList().size();
+        }
+
+        return drinkDao.getDrinkList()
+                .stream()
+                .map(DrinkDTO::drinkToDTO)
+                .collect(Collectors.toList())
+                .subList(fromIndex, toIndex);
+    }
+
+    @Transactional
+    public Set<String> getUniqueGlassesNameList() {
+        return drinkDao.getDrinkList()
+                .stream()
+                .map(Drink::getGlassType)
+                .collect(Collectors.toSet());
     }
 
     public List<Drink> findRecipeByCategoryId(List<Long> ids) {
